Changed service methods to ServiceChallengeStatistics
