+++ itachallenge-user/src/main/java/com/itachallenge/user/service/IServiceChallengeStatistics.java
@@ -8,4 +8,5 @@ import java.util.UUID;
 
 public interface IServiceChallengeStatistics {
     Mono<List<ChallengeStatisticsDto>> getChallengeStatistics(List<UUID> challengeIds);
+    Mono<Long> getBookmarkCountByIdChallenge(UUID idChallenge);
 }
